backend/requirements.txt
asn1crypto==0.24.0
boto==2.49.0
boto3==1.7.4
botocore==1.10.62
cached-property==1.4.3
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
comtypes==1.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3
dj-static==0.0.6
Django==2.0.5
django-buckets==0.1.25
django-cors-headers==2.3.0
django-environ==0.4.5
django-filter==1.1.0
django-rest-swagger==2.2.0
django-s3-upload==0.2
django-s3direct==1.0.4
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==2.1
docker==3.4.1
docker-compose==1.22.0
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
drf-yasg==1.11.0
drfdocs==0.0.11
h5py==2.8.0
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
lxml==4.2.4
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.5
openapi-codec==1.3.2
Pillow==5.2.0
preview-generator==0.2.3
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.18
PyJWT==1.5.2
PyPDF2==1.26.0
python-dateutil==2.7.3
python-docx==0.8.6
python-magic==0.4.15
python-magic-bin==0.4.14
pytz==2018.5
PyYAML==3.13
reportlab==3.5.4
requests==2.18.4
ruamel.yaml==0.15.77
s3transfer==0.1.13
scipy==1.1.0
simplejson==3.16.0
six==1.11.0
static3==0.7.0
text-to-image==0.0.5
texttable==0.9.1
unoconv==0.8.2
uritemplate==3.0.0
urllib3==1.22
Wand==0.4.4
websocket-client==0.48.0
Werkzeug==0.14.1
